Rename projects and other changes
