+++ CandidatesAndEmployees/CandidatesAndEmployees.csproj
@@ -6,14 +6,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="UserType.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="UserType.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Bogus" Version="34.0.1" />
   </ItemGroup>
