+++ CandidatesAndEmployees/CandidatesAndEmployees.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31321.278
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Users", "Users.csproj", "{6BD806CF-068B-4462-A796-86E2174E1AB3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CandidatesAndEmployees", "CandidatesAndEmployees.csproj", "{6BD806CF-068B-4462-A796-86E2174E1AB3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
